refactor: update auth imports to use authUtils across all components
