+++ apps/web/src/pages/AdminPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useAuth } from '../store/AuthContext';
+import { useAuth } from '../store/authUtils';
 import { Link } from 'react-router-dom';
 import { ROUTES } from '../routes';
