+++ apps/web/src/pages/DashboardPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { useAuth } from '../store/AuthContext';
+import { useAuth } from '../store/authUtils';
 import { useProfile } from '../hooks/useProfile';
 import { PATHS } from '../routes';
