+++ apps/web/src/pages/LoginPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
-import { useAuth } from '../store/AuthContext';
+import { useAuth } from '../store/authUtils';
 import { PATHS } from '../routes';
 
 /**
