+++ apps/web/src/routes/ProtectedRoute.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import { MemoryRouter, Route, Routes } from 'react-router-dom';
 import ProtectedRoute from './ProtectedRoute';
-import { useAuth } from '../store/AuthContext';
+import { useAuth } from '../store/authUtils';
 import { ROLES } from '../types/auth';
 
 // Mock the auth context
